% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_combined_data.R
\name{get_combined_data}
\alias{get_combined_data}
\title{Read ad combine USDA Organic Census tables}
\usage{
get_combined_data()
}
\value{
A data frame of USDA organic census data, to be saved as organic_census
}
\description{
Reads in PDF/Excel files from USDA generates a data frame. This function does not require any parameters
}
